The SMBClient.framework change, which required some parameters changes. Also correct a couple of warnings

+++ dcerpc/libdcethread/dcethread-private.c
@@ -260,7 +260,7 @@ dcethread__sanity(dcethread* thread)
         DCETHREAD_ERROR("NULL thread encountered");
         return;
     }
-    if ((long)thread->refs < 0)
+    if ((int)thread->refs < 0)
         DCETHREAD_ERROR("Thread %p: ref count < 0", thread);
     if (!thread->flag.locked)
         DCETHREAD_ERROR("Thread %p: not locked when expected", thread);
